agent: Extract cloudwatch specific tests
